# forge/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Forge Platform API",
    "description": "Complete API documentation for the Forge AI automation platform",
    "version": "1.0.0",
    "contact": {
      "name": "Forge Support",
      "email": "support@forge.ai"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "https://api.forge.ai/v1",
      "description": "Production Server"
    },
    {
      "url": "http://localhost:8000/api/v1",
      "description": "Development Server"
    }
  ],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "http",
      "scheme": "bearer",
      "bearerFormat": "JWT"
    },
    "apiKey": {
      "type": "apiKey",
      "in": "header",
      "name": "X-API-Key"
    }
  },
  "security": [
    {
      "bearerAuth": []
    },
    {
      "apiKey": []
    }
  ],
...
  "paths": {
...
    "/workflows": {
      "get": {
        "tags": ["Workflows"],
        "summary": "List all workflows",
        "security": [
          { "bearerAuth": [] }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "schema": { "type": "integer", "default": 0 }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "default": 20 }
          },
          {
            "name": "status",
            "in": "query",
            "schema": { "type": "string", "enum": ["active", "inactive", "archived"] }
          }
        ],
        "responses": {
          "200": {
            "description": "List of workflows",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "string" },
                          "name": { "type": "string" },
                          "description": { "type": "string" },
                          "status": { "type": "string" },
                          "createdAt": { "type": "string", "format": "date-time" }
                        }
                      }
                    },
                    "total": { "type": "integer" }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Workflows"],
        "summary": "Create a new workflow",
        "security": [
          { "bearerAuth": [] }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": { "type": "string" },
                  "description": { "type": "string" },
                  "steps": { "type": "array" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Workflow created"
          }
        }
      }
    },
    "/workflows/{id}": {
      "get": {
        "tags": ["Workflows"],
        "summary": "Get workflow details",
        "security": [
          { "bearerAuth": [] }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Workflow details"
          },
          "404": {
            "description": "Workflow not found"
          }
        }
      },
      "put": {
        "tags": ["Workflows"],
        "summary": "Update workflow",
        "security": [
          { "bearerAuth": [] }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Workflow updated"
          }
        }
      },
      "delete": {
        "tags": ["Workflows"],
        "summary": "Delete workflow",
        "security": [
          { "bearerAuth": [] }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "204": {
            "description": "Workflow deleted"
          }
        }
      }
    },
...
  }
}